requirements.txt
websockets>=12.0
requests>=2.31.0
urllib3>=2.0.0
python-socketio>=5.11.0
